changed: put MILU in separate compile unit
