+++ opm/simulators/linalg/FlowLinearSolverParameters.hpp
@@ -24,14 +24,10 @@
 #ifndef OPM_FLOWLINEARSOLVERPARAMETERS_HEADER_INCLUDED
 #define OPM_FLOWLINEARSOLVERPARAMETERS_HEADER_INCLUDED
 
-#include <opm/common/utility/parameters/ParameterGroup.hpp>
-#include <opm/simulators/linalg/ParallelOverlappingILU0.hpp>
+#include <opm/simulators/linalg/MILU.hpp>
 
 #include <opm/simulators/linalg/linalgproperties.hh>
 #include <opm/models/utils/parametersystem.hh>
-
-#include <array>
-#include <memory>
 
 namespace Opm {
 template <class TypeTag>
